+++ src/Part1/Main.java
@@ -1,3 +1,5 @@
+package Part1;
+
 public class Main {
     public static void main(String[] args) {
         System.out.println("Hello world!");
