+++ src/Part1/Pt1_10871.java
@@ -1,24 +1,22 @@
+package Part1;
+
 import java.io.*;
 
-public class Pt1_2588 {
+public class Pt1_10871 {
     public static void main(String[] args) throws IOException {
-
         BufferedReader bf = new BufferedReader(new InputStreamReader(System.in));
         BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(System.out));
 
         int N = Integer.parseInt(bf.readLine());
         int M = Integer.parseInt(bf.readLine());
 
-        for (int i=String.valueOf(M).length()-1 ; i>=0 ; i--){
-            char target;
-            target = String.valueOf(M).charAt(i);
 
-            bw.write((Character.getNumericValue(target) * N) +"\n");
+        for(int i=0; i<N ; i++){
+            int res = Integer.parseInt(bf.readLine());
+
+            if(res < M) bw.write(String.valueOf(res));
         }
 
-        bw.write(String.valueOf(N*M));
-
         bw.flush();
-
     }
 }
