+++ src/Part1/Pt1_2753.java
@@ -1,18 +1,18 @@
+package Part1;
+
 import java.io.*;
 
-public class Pt1_10871 {
+public class Pt1_2753 {
     public static void main(String[] args) throws IOException {
         BufferedReader bf = new BufferedReader(new InputStreamReader(System.in));
         BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(System.out));
 
         int N = Integer.parseInt(bf.readLine());
-        int M = Integer.parseInt(bf.readLine());
 
-
-        for(int i=0; i<N ; i++){
-            int res = Integer.parseInt(bf.readLine());
-
-            if(res < M) bw.write(String.valueOf(res));
+        if((N%4 == 0 && N%100 != 0) || N%400 == 0){
+            bw.write("1");
+        }else{
+            bw.write("0");
         }
 
         bw.flush();
